Clean up conversation forms a bit.
